auth-aws: remove E2E cluster setup (#12)
The auth-aws plugin brought an end-to-end test setup which spins up an
OpenBao cluster with the plugin enabled. This worked fine when part of
the main codebase, but an internal dependency of the tests now cause issues
when used in another repository.

Unfortunately the dependency in question (openbao/http) will cause
`go mod vendor` to throw an error because the required embeds for an UI
build can't be found. This in turn causes all dependabot runs to fail
and we currently don't receive any automated updates.

There is no easy way around this error, so as the simplest fix this
removes the E2E test cluster setup for the auth-aws plugin. Long term it
might makes sense to think about how we can setup E2E tests for all the
plugins in this repo.

Signed-off-by: Jan Martens <jan@martens.eu.org>