Fix plugin OCI image uploads (#25)
* fix plugin image naming

The individual images in our multi-arch plugin manifests were missing
the proper `ghcr.io/openbao` prefix. They are added by their SHA Hash to
the Manifest, so this did not impact functionality. It's just a cosmetic
change so you can see all the individual images with the correct tag if
you build locally.

Signed-off-by: Jan Martens <jan@martens.eu.org>

* push all Manifest images

Buildah versions <= 1.38.0 did not push all the images contained in a
multi-arch Manifest, only the manifest itself. The Ubuntu runners
we are using are shipping Buildah 1.33, so we need to explicitly set
`--all` when pushing the Manifest to ensure all images it contains are
pushed.

Signed-off-by: Jan Martens <jan@martens.eu.org>

---------

Signed-off-by: Jan Martens <jan@martens.eu.org>